+++ RELEASES.md
@@ -15,3 +15,11 @@ Based on:
 - Speakeasy CLI 1.598.2 (2.674.3) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v5.2.0-alpha.33] .
+
+## 2025-08-09 01:16:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.600.2 (2.677.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.34] .